+++ assembly/src/ast/nodes/mod.rs
@@ -268,7 +268,7 @@ pub enum Instruction {
 
     // ----- STARK proof verification -------------------------------------------------------------
     FriExt2Fold4,
-    RandCombMain,
+    RCombBase,
 
     // ----- exec / call --------------------------------------------------------------------------
     ExecLocal(u16),
@@ -526,8 +526,10 @@ impl fmt::Display for Instruction {
             Self::MTreeSet => write!(f, "mtree_set"),
             Self::MTreeMerge => write!(f, "mtree_merge"),
             Self::MTreeVerify => write!(f, "mtree_verify"),
+
+            // ----- STARK proof verification -----------------------------------------------------
             Self::FriExt2Fold4 => write!(f, "fri_ext2fold4"),
-            Self::RandCombMain => write!(f, "rcomb1"),
+            Self::RCombBase => write!(f, "rcomb_base"),
 
             // ----- exec / call ------------------------------------------------------------------
             Self::ExecLocal(index) => write!(f, "exec.{index}"),