+++ processor/src/operations/mod.rs
@@ -149,7 +149,7 @@ where
             Operation::MpVerify => self.op_mpverify()?,
             Operation::MrUpdate => self.op_mrupdate()?,
             Operation::FriE2F4 => self.op_fri_ext2fold4()?,
-            Operation::RanComb1 => self.op_rcomb_base()?,
+            Operation::RCombBase => self.op_rcomb_base()?,
         }
 
         self.advance_clock()?;